Move utilities to manipulate requests and responses
These utilities do not belong in `TransportLayer`.
