+++ docs/utils.rst
@@ -1,0 +1,5 @@
+Utilities
+=========
+
+.. autoclass:: sipua.create_response
+   :members:
