+++ src/sipua/__init__.py
@@ -11,8 +11,10 @@ from .transaction import (
     TransactionLayer,
 )
 from .transport import TransportAddress, TransportLayer
+from .utils import create_response
 
 __all__ = [
+    "create_response",
     "ServerInviteTransaction",
     "ServerNonInviteTransaction",
     "TransactionLayer",
