+++ tests/test_transport.py
@@ -16,14 +16,13 @@ import websockets.asyncio.client
 import websockets.asyncio.connection
 
 from sipua.transport import (
-    ANY_HOST,
-    ANY_TRANSPORT,
     WEBSOCKET_SUBPROTOCOL,
     TransportAddress,
     TransportLayer,
     get_transport_destination,
     update_request_via,
 )
+from sipua.utils import create_ack, create_contact, create_response, create_via
 
 from .utils import asynctest, lf2crlf, parse_request
 
@@ -39,13 +38,8 @@ def create_request(
             uri, parameters=sipmessage.Parameters(transport=uri_transport)
         )
     request = sipmessage.Request(method="OPTIONS", uri=uri)
-    request.via = [
-        sipmessage.Via(
-            transport=ANY_TRANSPORT,
-            host=ANY_HOST,
-            parameters=sipmessage.Parameters(branch="z9hG4bK1e5b2b763d"),
-        )
-    ]
+    with patch("sipua.utils.random_string", new=lambda x: "1e5b2b763d"):
+        request.via = [create_via()]
     request.max_forwards = 70
     request.to_address = sipmessage.Address.parse("sip:bob@example.com")
     request.from_address = sipmessage.Address.parse(
@@ -53,14 +47,7 @@ def create_request(
     )
     request.call_id = "126a8db08eba7fb6"
     request.cseq = sipmessage.CSeq(1, "OPTIONS")
-    request.contact = [
-        sipmessage.Address(
-            uri=sipmessage.URI(
-                scheme="sip",
-                host=ANY_HOST,
-            )
-        )
-    ]
+    request.contact = [create_contact()]
     return request
 
 
@@ -230,7 +217,7 @@ class TcpTransportChannelTest(BaseTestCase):
             )
 
             # Send response.
-            response = transport.create_response(request=request, code=200)
+            response = create_response(request=request, code=200)
             is_reliable = await transport.send_message(response)
             self.assertTrue(is_reliable)
 
@@ -358,7 +345,7 @@ class UdpTransportChannelTest(BaseTestCase):
             )
 
             # Send response.
-            response = transport.create_response(request=request, code=200)
+            response = create_response(request=request, code=200)
             is_reliable = await transport.send_message(response)
             self.assertFalse(is_reliable)
 
@@ -480,7 +467,7 @@ class WebsocketTransportChannelTest(BaseTestCase):
             )
 
             # Send response.
-            response = transport.create_response(request=request, code=200)
+            response = create_response(request=request, code=200)
             is_reliable = await transport.send_message(response)
             self.assertTrue(is_reliable)
 
@@ -560,7 +547,7 @@ a=fmtp:101 0-15
 a=silenceSupp:off - - - -
 """)
 
-        response = transport.create_response(request=request, code=200)
+        response = create_response(request=request, code=200)
         self.assertMessage(
             response,
             """SIP/2.0 200 OK
@@ -573,7 +560,7 @@ CSeq: 1 INVITE
 """,
         )
 
-        ack = transport.create_ack(request=request, response=response)
+        ack = create_ack(request=request, response=response)
         self.assertMessage(
             ack,
             """ACK sip:+33233445566@127.0.0.1:5060 SIP/2.0
@@ -613,7 +600,7 @@ a=fmtp:101 0-15
 a=silenceSupp:off - - - -
 
 """)
-        response = transport.create_response(request=request, code=200)
+        response = create_response(request=request, code=200)
         self.assertMessage(
             response,
             """SIP/2.0 200 OK
